feat(interface-types) Extract `InterfaceType` and `RecordType` into its own `types` module.
