+++ lib/interface-types/src/interpreter/instructions/mod.rs
@@ -164,11 +164,9 @@ where
 #[cfg(test)]
 pub(crate) mod tests {
     use crate::{
-        ast,
-        interpreter::wasm::{
-            self,
-            values::{InterfaceType, InterfaceValue},
-        },
+        ast::*,
+        interpreter::wasm::{self, values::InterfaceValue},
+        types::*,
     };
     use std::{cell::Cell, collections::HashMap, convert::TryInto, ops::Deref, rc::Rc};
 
@@ -265,7 +263,7 @@ pub(crate) mod tests {
         pub(crate) exports: HashMap<String, Export>,
         pub(crate) locals_or_imports: HashMap<usize, LocalImport>,
         pub(crate) memory: Memory,
-        pub(crate) wit_types: Vec<ast::Type>,
+        pub(crate) wit_types: Vec<Type>,
     }
 
     impl Instance {
@@ -322,10 +320,10 @@ pub(crate) mod tests {
                     hashmap
                 },
                 memory: Memory::new(vec![Cell::new(0); 128]),
-                wit_types: vec![ast::Type::Record(ast::RecordType {
+                wit_types: vec![Type::Record(RecordType {
                     fields: vec1![
                         InterfaceType::I32,
-                        InterfaceType::Record(ast::RecordType {
+                        InterfaceType::Record(RecordType {
                             fields: vec1![InterfaceType::String, InterfaceType::F32],
                         }),
                         InterfaceType::I64,
@@ -351,7 +349,7 @@ pub(crate) mod tests {
             Some(&self.memory)
         }
 
-        fn wit_type(&self, index: u32) -> Option<&ast::Type> {
+        fn wit_type(&self, index: u32) -> Option<&Type> {
             self.wit_types.get(index as usize)
         }
     }
